+++ src/pages/about.ts
@@ -1,9 +1,9 @@
 
-import { Layout, Paragraph } from "../../Lib/+innerscope";
+import { LinearLayout, Paragraph } from "../../Lib/+innerscope";
 
 export default function AboutPage() {
-    const layout = Layout("linear", document.body);
-    layout.childAlignment("center");
-
-    Paragraph(`You Are On The Get Started Page`, layout);
+    const page = LinearLayout(document.body);
+    page.ElementAlignment = "CENTER"
+    page.ParentFill = "FILLXY"
+    Paragraph(`You Are On The Get Started Page`, page);
 }